vutil.c: Simplify locale handling when using POSIX 2008
I read the code over and realized that there was a much simpler way to
do things when using POSIX 2008 locales.  Instead of all the rigamorole
about saving/restoring locales, just toggle to the C object using the
thread-safe uselocale() libc function.  This trivally gets us to the C
locale, and hence a dot radix.