locale.c: Change parameter specification
S_populate_hash_from_localeconv() takes a mask of one or two bits that
specify which of LC_NUMERIC and/or LC_MONETARY to operate on.  Prior to
this commit, the appropriate bit was based on our internal index for the
relevant category(ies).  But we already have #defines of 0 for one, and
1 for the other.  This commit changes to base the index on that.  This
will allow simplification in the next commit, and going forward.